Raise PHPStan to level 7
Tightens types across DB-row return shapes (list<> for fetchAll),
adds class-string typing to HandlerRegistry, narrows callable arity
detection, and validates ReflectionClass/Method inputs. Restructures
StateMachineBuilder partial state updates so PHPStan tracks the shape
correctly, and reshapes Queue::stats() assembly to preserve the
declared array shape.

+++ src/MysqliPdoStatement.php
@@ -189,7 +189,7 @@ class MysqliPdoStatement extends \PDOStatement {
 			while ( $row = mysqli_fetch_assoc( $result ) ) {
 				$this->rows[] = $row;
 			}
-			$this->row_count = mysqli_num_rows( $result );
+			$this->row_count = (int) mysqli_num_rows( $result );
 			mysqli_free_result( $result );
 			return;
 		}
@@ -209,7 +209,7 @@ class MysqliPdoStatement extends \PDOStatement {
 			while ( $row = mysqli_fetch_assoc( $result ) ) {
 				$this->rows[] = $row;
 			}
-			$this->row_count = mysqli_num_rows( $result );
+			$this->row_count = (int) mysqli_num_rows( $result );
 			mysqli_free_result( $result );
 			return;
 		}
